+++ src/main/Settings.js
@@ -4,8 +4,8 @@ import path from 'path'
 export default {
   docPath: path.join(app.getPath('documents'), 'Cartaro'),
   window: {
-    width: 1280,
-    height: 1000,
+    width: 1680,
+    height: 1250,
     minWidth: 900,
     minHeight: 600
   },